+++ CHANGES.md
@@ -20,6 +20,7 @@ Version 1.1.0 (pending)
 - Don't require phone or email for Child badges
 - Added required note to Attendee Edit screen, saves to Attendee History
 - Added "Ask Me My Pronouns" to available preferred pronouns
+- Allow fewer digits in check number field (at-con payment)
 
 Version 1.0.1 (11/14/2019)
 - Staff check in: don't search staff list until 2 characters have been entered in search (for performance)